Objects are all created

+++ Towers-of-Hanoi-2/src/objects/Point.java
@@ -3,10 +3,10 @@ package objects;
 import java.util.Objects;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
